apps/suap_id/requirements-app.txt
-e ../ege_django_theme

python-social-auth==0.3.6
social-auth-app-django==2.1.0